test: reorganize reservation and theme controller tests
