+++ src/test/java/roomescape/controller/ThemeControllerTest.java
@@ -17,50 +17,17 @@ import org.springframework.test.context.jdbc.Sql;
 @DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
 public class ThemeControllerTest {
 
-    private Map<String, Object> themeParams() {
-        Map<String, Object> params = new HashMap<>();
-        params.put("name", "이든의 공포 하우스");
-        params.put("description", "이든이 귀신으로 나옴");
-        params.put("imgUrl", "링크~");
-        return params;
-    }
-
-    private Map<String, Object> reservationParams() {
-        Map<String, Object> params = new HashMap<>();
-        params.put("name", "브라운");
-        params.put("date", "2023-08-05");
-        params.put("timeId", 1);
-        params.put("themeId", 1);
-        return params;
+    @Test
+    void 테마_조회() {
+        RestAssured.given().log().all()
+                .when().get("/api/v1/themes")
+                .then().log().all()
+                .statusCode(200)
+                .body("size()", is(0));
     }
 
     @Test
     void 테마_추가() {
-        Map<String, Object> adminThemeParams = themeParams();
-        adminThemeParams.put("userName", "ADMIN");
-        RestAssured.given().log().all()
-                .contentType(ContentType.JSON)
-                .body(adminThemeParams)
-                .when().post("/api/v1/themes")
-                .then().log().all()
-                .statusCode(201)
-                .body("id", is(1));
-    }
-
-    @Test
-    void 테마_추가시_관리자가_아닌경우_401을_반환한다() {
-        Map<String, Object> userThemeParams = themeParams();
-        userThemeParams.put("userName", "정콩이");
-        RestAssured.given().log().all()
-                .contentType(ContentType.JSON)
-                .body(userThemeParams)
-                .when().post("/api/v1/themes")
-                .then().log().all()
-                .statusCode(401);
-    }
-
-    @Test
-    void 테마_조회() {
         Map<String, Object> adminThemeParams = themeParams();
         adminThemeParams.put("userName", "ADMIN");
         RestAssured.given().log().all()
@@ -87,7 +54,14 @@ public class ThemeControllerTest {
                 .body(adminThemeParams)
                 .when().post("/api/v1/themes")
                 .then().log().all()
-                .statusCode(201);
+                .statusCode(201)
+                .body("id", is(1));
+
+        RestAssured.given().log().all()
+                .when().get("/api/v1/themes")
+                .then().log().all()
+                .statusCode(200)
+                .body("size()", is(1));
 
         RestAssured.given().log().all()
                 .contentType(ContentType.JSON)
@@ -95,6 +69,24 @@ public class ThemeControllerTest {
                 .when().delete("/api/v1/themes/1")
                 .then().log().all()
                 .statusCode(204);
+
+        RestAssured.given().log().all()
+                .when().get("/api/v1/themes")
+                .then().log().all()
+                .statusCode(200)
+                .body("size()", is(0));
+    }
+
+    @Test
+    void 테마_추가시_관리자가_아닌경우_401을_반환한다() {
+        Map<String, Object> userThemeParams = themeParams();
+        userThemeParams.put("userName", "정콩이");
+        RestAssured.given().log().all()
+                .contentType(ContentType.JSON)
+                .body(userThemeParams)
+                .when().post("/api/v1/themes")
+                .then().log().all()
+                .statusCode(401);
     }
 
     @Test
@@ -156,5 +148,22 @@ public class ThemeControllerTest {
                 .then().log().all()
                 .statusCode(200)
                 .body("size()", is(10));
+    }
+
+    private Map<String, Object> themeParams() {
+        Map<String, Object> params = new HashMap<>();
+        params.put("name", "이든의 공포 하우스");
+        params.put("description", "이든이 귀신으로 나옴");
+        params.put("imgUrl", "링크~");
+        return params;
+    }
+
+    private Map<String, Object> reservationParams() {
+        Map<String, Object> params = new HashMap<>();
+        params.put("name", "브라운");
+        params.put("date", "2023-08-05");
+        params.put("timeId", 1);
+        params.put("themeId", 1);
+        return params;
     }
 }
